% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_Gram_Schmidt}
\alias{apply_Gram_Schmidt}
\title{Function: Gram-Schmidt}
\usage{
apply_Gram_Schmidt(A)
}
\arguments{
\item{A}{matrix of real numbers, where the set of vectors is represented by each row of the matrix.}
}
\value{
Returns a matrix, where each row corresponds to the vector normalized by the Gram-Schimit process.
}
\description{
This functions performs the orthonormalization the set of vectors using the Gram-Schmidt process.
}
\examples{
A <- matrix(sample.int(4, 9, replace = TRUE),  nrow = 3)
print(A)
apply_Gram_Schmidt(A)
}
