% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_pseudo_inverse}
\alias{apply_pseudo_inverse}
\title{Function: Pseudo Inverse}
\usage{
apply_pseudo_inverse(A)
}
\arguments{
\item{A}{matrix of real numbers}
}
\value{
Returns a matrix \code{A} representing the pseudo-inverse of A.
}
\description{
Given an m×n real matrix A, this function calculates the pseudo-inverse A+ of the matrix A.
}
\examples{
A <- matrix(c(1, 0, 1,
              0, 1, 1), ncol = 2)
print(apply_pseudo_inverse(A))
}
