% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cholesky}
\alias{cholesky}
\title{Function: Cholesky}
\usage{
cholesky(A)
}
\arguments{
\item{A}{matrix of real numbers}
}
\value{
Returns a matrix \code{A} representing the cholesky decomposition
}
\description{
Given an m×n real matrix A, this function calculates the cholesky of matrix A.
}
\examples{
A <- matrix(c(4.0,-1.0,0.0,0.0,-1.0,4.0,-1.0,0.0,0.0,-1.0,4.0,-1.0,0.0,0.0,-1.0,4.0),nrow=4,ncol=4)
print(cholesky(A))
}
