% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{dotprod}
\alias{dotprod}
\title{Vector Dot Function}
\usage{
dotprod(u, v)
}
\arguments{
\item{u}{vector of real numbers.}

\item{v}{vector of real numbers.}
}
\value{
Returns a dot product between \code{u} and \code{v}.
}
\description{
Find vector dot product between two vectors.
}
\examples{
u <- c(-1, -2, 3)
v <- c(4, 0, -8)
dotprod(u, v)
}
