% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_orthogonal_complement}
\alias{find_orthogonal_complement}
\title{Function: Find Orthogonal Complement}
\usage{
find_orthogonal_complement(A)
}
\arguments{
\item{A}{matrix of real numbers, where each line corresponds to each vector}
}
\value{
Returns a matrix \code{A} containing the orthogonal complement of the set of vectors.
}
\description{
Compute the orthogonal complement of a set of vectors {v1, v2, ...}
}
\examples{
A <- matrix(c(1, 2, 3, 4,
              2, 5, 0, 1), byrow = TRUE, nrow = 2)
print(find_orthogonal_complement(A))
B <- matrix(c(3, 12, -1), byrow = TRUE, nrow = 1)
print(find_orthogonal_complement(B))
}
