% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{magnitude}
\alias{magnitude}
\title{Vector Magnitude Function}
\usage{
magnitude(u)
}
\arguments{
\item{u}{vector of real numbers.}
}
\value{
Returns a magnitude (length) of \code{u}.
}
\description{
Find the vector magnitude (length)
}
\examples{
u <- c(1,0,-3)
magnitude(u)
u <- sample(5, 4)
magnitude(u)
}
