% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{projection}
\alias{projection}
\title{Vector Projection Function}
\usage{
projection(v, u)
}
\arguments{
\item{v}{vector of real numbers.}

\item{u}{vector of real numbers.}
}
\value{
Returns a projection of \code{u} onto \code{v}.
}
\description{
Find the vector projection.
}
\examples{
v <- c(1, 2)
u <- c(3, -8)
projection(v, u)
v <- c(1, 0, 3)
u <- c(-1, 4, 2)
projection(v, u)
}
