% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{unitvec}
\alias{unitvec}
\title{Vector Unit Function}
\usage{
unitvec(u)
}
\arguments{
\item{u}{vector of real numbers.}
}
\value{
Returns a unit vector of \code{u}.
}
\description{
Find the unit vector.
}
\examples{
u <- c(1,0,-3)
unitvec(u)
u <- sample(5, 4)
unitvec(u)
}
